Moved datadata related helpers , and fixed output folder for agg files
